Consolidate UseTrusteeFormValidation into UseTrusteeForm
Jira ticket: CAMS-584
