+++ user-interface/src/trustees/UseTrusteeForm.tsx
@@ -1,7 +1,32 @@
 import { useState } from 'react';
 import { ContactInformation } from '@common/cams/contact';
-import { trusteeFormDataSpec } from './UseTrusteeFormValidation';
-import { ChapterType, TrusteeInput, TrusteeStatus } from '@common/cams/trustees';
+import {
+  ChapterType,
+  TRUSTEE_STATUS_VALUES,
+  TrusteeInput,
+  TrusteeStatus,
+} from '@common/cams/trustees';
+import {
+  flattenReasonMap,
+  validateEach,
+  validateObject,
+  ValidationSpec,
+} from '@common/cams/validation';
+import V from '@common/cams/validators';
+import { EMAIL_REGEX, EXTENSION_REGEX, PHONE_REGEX, ZIP_REGEX } from '@common/cams/regex';
+
+export const TRUSTEE_SPEC: Readonly<ValidationSpec<TrusteeFormData>> = {
+  name: [V.minLength(1, 'Trustee name is required')],
+  address1: [V.minLength(1, 'Address is required')],
+  address2: [V.optional(V.maxLength(50))],
+  city: [V.minLength(1, 'City is required')],
+  state: [V.exactLength(2, 'State is required')],
+  zipCode: [V.matches(ZIP_REGEX, 'ZIP code must be 5 digits or 9 digits with a hyphen')],
+  email: [V.matches(EMAIL_REGEX, 'Email must be a valid email address')],
+  phone: [V.matches(PHONE_REGEX, 'Phone must be a valid phone number')],
+  extension: [V.optional(V.matches(EXTENSION_REGEX, 'Extension must be 1 to 6 digits'))],
+  status: [V.isInSet<TrusteeStatus>([...TRUSTEE_STATUS_VALUES])],
+};
 
 export interface TrusteeFormData {
   name: string;
@@ -29,13 +54,13 @@ export interface TrusteeFormValidation {
   validateFieldAndUpdate: (
     field: keyof TrusteeFormData,
     value: string,
-    spec: Partial<typeof trusteeFormDataSpec>,
+    spec: Partial<typeof TRUSTEE_SPEC>,
   ) => string | null;
   clearErrors: () => void;
   clearFieldError: (field: string) => void;
   isFormValidAndComplete: (
     formData: TrusteeFormData,
-    spec: Partial<typeof trusteeFormDataSpec>,
+    spec: Partial<typeof TRUSTEE_SPEC>,
   ) => boolean;
 }
 
@@ -50,6 +75,30 @@ export type TrusteeFormState = {
 type UseTrusteeFormProps = {
   initialState: TrusteeFormState;
 };
+
+/**
+ * Validates individual form fields with specific business rules
+ */
+function validateField(
+  field: keyof TrusteeFormData,
+  value: string,
+  spec: Partial<typeof TRUSTEE_SPEC>,
+): string | null {
+  // Convert to string and trim
+  const stringValue = String(value);
+  const trimmedValue = stringValue.trim();
+
+  if (field === 'status' || (field === 'extension' && !trimmedValue)) {
+    return null;
+  }
+
+  if (spec?.[field]) {
+    const result = validateEach(spec[field], trimmedValue);
+    return result.valid ? null : result.reasons!.join(' ');
+  } else {
+    return null;
+  }
+}
 
 export function useTrusteeForm({ initialState }: UseTrusteeFormProps) {
   const getInitialFormData = (): TrusteeFormData => {
@@ -82,6 +131,7 @@ export function useTrusteeForm({ initialState }: UseTrusteeFormProps) {
   };
 
   const [formData, setFormData] = useState<TrusteeFormData>(getInitialFormData());
+  const [fieldErrors, setFieldErrors] = useState<Record<string, string>>({});
 
   const updateField = (field: keyof TrusteeFormData, value: unknown) => {
     setFormData((prev) => ({ ...prev, [field]: value }));
@@ -117,11 +167,77 @@ export function useTrusteeForm({ initialState }: UseTrusteeFormProps) {
     return trimmedData;
   };
 
+  /**
+   * Validates a single field and updates the field errors state
+   */
+  const validateFieldAndUpdate = (
+    field: keyof TrusteeFormData,
+    value: string,
+    spec: Partial<typeof TRUSTEE_SPEC>,
+  ): string | null => {
+    const error = validateField(field, value, spec);
+
+    setFieldErrors((prevErrors) => {
+      if (error) {
+        return { ...prevErrors, [field]: error };
+      } else {
+        const { [field]: _, ...rest } = prevErrors;
+        return rest;
+      }
+    });
+
+    return error;
+  };
+
+  /**
+   * Clears all validation errors
+   */
+  const clearErrors = (): void => {
+    setFieldErrors({});
+  };
+
+  /**
+   * Clears validation error for a specific field
+   */
+  const clearFieldError = (field: string): void => {
+    setFieldErrors((prevErrors) => {
+      const { [field]: _, ...rest } = prevErrors;
+      return rest;
+    });
+  };
+
+  /**
+   * Checks if form is both valid (no errors) and complete (all required fields filled)
+   */
+  const isFormValidAndComplete = (
+    formData: TrusteeFormData,
+    spec: Partial<typeof TRUSTEE_SPEC>,
+  ): boolean => {
+    const results = validateObject(spec, formData);
+    if (!results.valid && results.reasonMap) {
+      const newFieldErrors = Object.fromEntries(
+        Object.entries(flattenReasonMap(results.reasonMap)).map(([jsonPath, reasons]) => {
+          const field = jsonPath.split('.')[1];
+          return [field, reasons.join('. ') + '.'];
+        }),
+      );
+      setFieldErrors(newFieldErrors);
+    }
+    return !!results.valid;
+  };
+
   return {
     formData,
     updateField,
     updateMultipleFields,
     resetForm,
     getFormData,
+
+    fieldErrors,
+    errors: Object.entries(fieldErrors).map(([field, message]) => ({ field, message })),
+    validateFieldAndUpdate,
+    clearErrors,
+    clearFieldError,
+    isFormValidAndComplete,
   };
 }
